moved store fixtures
